requirements.txt

pandas>=2.0.0
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
lxml>=4.9.0
openpyxl>=3.1.0
